refactor(download): split index.ts into cohesive sibling modules (part of #5165)
